blkdev: allow for 16bit block counts
On a 32bit system we can end up writing more than 256 blocks in a run when
we swap so blkdev needs to use 16bits for the block counter.

(We could make this a typedef for 8 v 16bit but it's not clear it matters
 enough)

+++ Kernel/dev/blkdev.h
@@ -27,7 +27,7 @@ struct blkparam {
     uint8_t swap_page;                      /* page to pass to map_swap */
     blkdev_t *blkdev;                       /* active block device */
     uint32_t lba;                           /* LBA for first sectors to transfer */
-    uint8_t nblock;                         /* number of sectors to transfer */
+    uint16_t nblock;                        /* number of sectors to transfer */
     bool is_read;                           /* true: read sectors, false: write sectors */
 };
 /* macros that inline assembler code can use to access blkparam fields */
